requirements.txt
streamlit
langchain
openai
python-dotenv
tiktoken
